Fix sentinel issue with detection of downed master
It turns out that sentinel will not check for a down master if the connect method returns an error. For non-ssl this is OK, as it's a non-blocking connect. A non-blocking connect will return EINPROGRESS if the other side is not available, which the redis code ignores if nonblocking. It sets up the connection, and continues through the main processing loop.  For SSL, though, the connect and negotiation portion itself is not non-blocking (sort-of) in hiredis, and it does return an error if the other side is not available. In here, we need to mask this for SSL, so the checks can continue properly.